Refactor: Update EmbeddingAttrs to ensure input tensor is of type long. Introduce ensemble_predictions_and_targets function for aggregating predictions and targets in loss and metrics based on ensemble indices. Enhance loss calculations in SimpleLoss and its subclasses to handle NaN values and improve metric selection in Trainer class.

+++ geqtrain/nn/_node.py
@@ -130,7 +130,7 @@ class EmbeddingAttrs(GraphModuleMixin, torch.nn.Module):
             if self.use_masking and attribute_name in self.fields_to_mask:
                 x = apply_masking(x, mask_token_index=emb_layer.weight.shape[0] -1) # last index is reserved for masking, make sure to match this in yaml
 
-            x_emb = emb_layer(x)
+            x_emb = emb_layer(x.long())
             out.append(x_emb)
 
         if self._has_numericals:
